+++ src/main/java/chess/domain/BoardFactory.java
@@ -1,4 +1,4 @@
-package chess.domain.board;
+package chess.domain;
 
 import static java.util.stream.Collectors.toList;
 import static java.util.stream.Collectors.toMap;
@@ -12,6 +12,9 @@ import chess.domain.piece.Pawn;
 import chess.domain.piece.Piece;
 import chess.domain.piece.Queen;
 import chess.domain.piece.Rook;
+import chess.domain.position.File;
+import chess.domain.position.Position;
+import chess.domain.position.Rank;
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
