+++ src/main/java/chess/domain/piece/King.java
@@ -2,7 +2,7 @@ package chess.domain.piece;
 
 import static chess.domain.piece.PieceType.KING;
 
-import chess.domain.board.Position;
+import chess.domain.position.Position;
 
 public class King extends Piece {
