+++ src/main/java/chess/domain/piece/Piece.java
@@ -1,6 +1,6 @@
 package chess.domain.piece;
 
-import chess.domain.board.Position;
+import chess.domain.position.Position;
 
 public abstract class Piece {
