+++ src/main/java/chess/domain/piece/Queen.java
@@ -2,7 +2,7 @@ package chess.domain.piece;
 
 import static chess.domain.piece.PieceType.QUEEN;
 
-import chess.domain.board.Position;
+import chess.domain.position.Position;
 
 public class Queen extends Piece {
