+++ src/main/java/chess/domain/position/Position.java
@@ -1,4 +1,4 @@
-package chess.domain.board;
+package chess.domain.position;
 
 import static java.util.stream.Collectors.toList;
 import static java.util.stream.Collectors.toMap;
