+++ src/main/java/chess/view/OutputView.java
@@ -2,9 +2,9 @@ package chess.view;
 
 import static java.util.stream.Collectors.toList;
 
-import chess.domain.board.File;
-import chess.domain.board.Position;
-import chess.domain.board.Rank;
+import chess.domain.position.File;
+import chess.domain.position.Position;
+import chess.domain.position.Rank;
 import chess.domain.piece.Color;
 import chess.domain.piece.Piece;
 import chess.domain.piece.PieceType;
