+++ src/test/java/chess/domain/BoardFactoryTest.java
@@ -1,4 +1,4 @@
-package chess.domain.board;
+package chess.domain;
 
 import static chess.domain.piece.PieceType.BISHOP;
 import static chess.domain.piece.PieceType.EMPTY;
@@ -9,8 +9,13 @@ import static chess.domain.piece.PieceType.QUEEN;
 import static chess.domain.piece.PieceType.ROOK;
 import static org.assertj.core.api.Assertions.assertThat;
 
+import chess.domain.Board;
+import chess.domain.BoardFactory;
 import chess.domain.piece.Piece;
 import chess.domain.piece.PieceType;
+import chess.domain.position.File;
+import chess.domain.position.Position;
+import chess.domain.position.Rank;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
